+++ java_ojs/src/models/ArticleFile.java
@@ -1,22 +1,22 @@
 package models;
 
 public class ArticleFile {
-	private int id;
-	private int article_id;
+	private long id;
+	private long article_id;
 	private String path;
-	public ArticleFile(int file_id,int article_id, String path) {
+	public ArticleFile(long file_id,long article_id, String path) {
 	
 		this.id = file_id;
 		this.article_id = article_id;
 		this.path = path;
 	}
-	public int getId() {
+	public long getId() {
 		return id;
 	}
-	public int getArticle_id() {
+	public long getArticle_id() {
 		return article_id;
 	}
-	public void setArticle_id(int article_id) {
+	public void setArticle_id(long article_id) {
 		this.article_id = article_id;
 	}
 	public String getPath() {
