+++ java_ojs/src/models/Metadata.java
@@ -1,12 +1,12 @@
 package models;
 
 public class Metadata {
-	private int id;
-	private int article_id;
+	private long id;
+	private long article_id;
 	private String competing_interests="";
 	private String funding="";
 	
-	public Metadata(int id, int article_id, String competing_interests, String funding) {
+	public Metadata(long id, long article_id, String competing_interests, String funding) {
 		super();
 		this.id = id;
 		this.article_id = article_id;
@@ -14,11 +14,11 @@ public class Metadata {
 		this.funding = funding;
 	}
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public int getArticle_id() {
+	public long getArticle_id() {
 		return article_id;
 	}
 	public void setArticle_id(int article_id) {
